Fix: исправлено зацикливание классов путём рефакторинга и выделения класса BotHelper
